+++ vfat.c
@@ -196,7 +196,7 @@ int clear_vses(Stream_t *Dir, int entrySlot, size_t last)
 		exit(1);
 	}
 	addFreeEntry(cache, entry.entry, last);
-	for (; entry.entry < last; ++entry.entry) {
+	for (; entry.entry < (signed int) last; ++entry.entry) {
 #if DEBUG
 		fprintf(stderr,"Clearing entry %d.\n", entry.entry);
 #endif
@@ -645,12 +645,12 @@ static void accountFreeSlots(struct scan_state *ssp, dirCacheEntry_t *dce)
  * normal vfat_lookup
  */
 int lookupForInsert(Stream_t *Dir,
-					char *dosname,
-					char *longname,
-					struct scan_state *ssp, 
-					int ignore_entry,
-					int source_entry,
-					int pessimisticShortRename)
+		    char *dosname,
+		    char *longname,
+		    struct scan_state *ssp, 
+		    int ignore_entry,
+		    int source_entry,
+		    int pessimisticShortRename)
 {
 	direntry_t entry;
 	int ignore_match;
@@ -698,13 +698,13 @@ int lookupForInsert(Stream_t *Dir,
 				break;
 			case DCET_USED:
 				if(!(dce->dir.attr & 0x8) &&
-				   dce->endSlot - 1 == source_entry)
+				   (signed int)dce->endSlot-1 == source_entry)
 				   accountFreeSlots(ssp, dce);
 
 				/* labels never match, neither does the 
 				 * ignored entry */
 				if( (dce->dir.attr & 0x8) ||
-				    (dce->endSlot - 1 == ignore_entry) )
+				    ((signed int)dce->endSlot-1==ignore_entry))
 					break;
 
 				/* check long name */
